Factor out problem creating and serializing for future schema changes with hopefully less searching.
